scrapers/requirements.txt
astroid==3.3.9
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
dill==0.3.9
dnspython==2.7.0
h11==0.14.0
idna==3.10
isort==6.0.1
lxml==5.3.1
mccabe==0.7.0
outcome==1.3.0.post0
platformdirs==4.3.7
pycparser==2.22
pymongo==4.11.3
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.30.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tomlkit==0.13.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
